# src/main/java/com/nuevospa/gestiontareas/dtos/TaskResponse.java
package com.nuevospa.gestiontareas.dtos;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.util.Date;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class TaskResponse {
    private Long id;
    private String title;
    private String description;
    private String status;
    private Long userId;
    private String username;
    private Date creationDate;
    private Date updateDate;

}